Mess with some other stuff before realizing error
Turns out the dimensions on our block data was wrong, which is why the
maze was rendering funny. Everything appears to be in a working state at
this point. Coming up, consider turning the maze into a custom primitive
rather than having the component attach to a generic entity.

The very next thing we're doing is figuring out how to scale things.