A very basic idea of scoring
This was mostly just so that we could register collision events. Rather
than registering it on the platforms, registering it on the Thruster.
This is probably not the way to do this, since the platform itself will
know if it's been 'hit' or not already, and will update the score
accordingly.

That of course leads to the realization that the score might need to
exist inside a separate, more "global" entity.